pyproject.toml
[project]
name = "supercoach-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.140.0",
    "crewai-tools>=0.49.0",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "postgrest>=1.1.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "supabase>=2.16.0",
    "tiktoken>=0.9.0",
    "typing-extensions>=4.14.1",
]
